% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabselgraphical.R
\name{getLamPath}
\alias{getLamPath}
\title{Create a regularization path - copied from pulsar.}
\usage{
getLamPath(max, min, len, log = FALSE)
}
\arguments{
\item{max}{maximum value for regularization (lambda)}

\item{min}{min value for lambda}

\item{len}{length of path}

\item{log}{log spacing}
}
\description{
Create a regularization path - copied from pulsar.
}

