% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabselgraphical.R
\name{quic.graphical_model}
\alias{quic.graphical_model}
\title{stability selection fit function for sparse inverse covariance using QUIC}
\usage{
quic.graphical_model(x, y, q, ...)
}
\arguments{
\item{x}{data matrix}

\item{y}{data matrix}

\item{q}{number of variables}

\item{...}{extra arguments - currently only lams (optional) a lambda path}
}
\value{
list with selected variables and path.
}
\description{
stability selection fit function for sparse inverse covariance using QUIC
}
\details{
This is a wrapper for QUIC to be used in stability selection. Pass it as
the fitfunction to stabsel. If you supply your own graphical model fitter, please
set the class to "graphical_model".
}
\examples{
if (require(huge)) {
 set.seed(10010)
 dat.hubs <- huge.generator(n=1000, d=40, graph="hub")
 pcutoff <- 0.75
 PFER <- 10
 s.hubs <- stabsel(x=dat.hubs$data, fitfun=quic.graphical_model, 
                  cutoff=pcutoff, PFER=PFER)

}
}
\seealso{
stabsel
}

