% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabselgraphical.R
\name{stabs.quic}
\alias{stabs.quic}
\title{stability selection fit function for sparse inverse covariance using QUIC}
\usage{
stabs.quic(x, y, q, ...)
}
\arguments{
\item{x}{data matrix}

\item{y}{data matrix}

\item{q}{number of variables}
}
\value{
list with selected variables and path.
}
\description{
stability selection fit function for sparse inverse covariance using QUIC
}
\details{
This is a wrapper for QUIC to be used in stability selection. Pass it as
the fitfunction to stabsel
}
\examples{
if (require(huge)) {
 set.seed(10010)
 dat.hubs <- huge.generator(n=N, d=40, graph="hub")
 pcutoff <- 0.75
 PFER <- 10
 s.hubs <- stabsel(x=dat.hubs$data, fitfun=stabs.quic, 
                  cutoff=pcutoff, PFER=PFER)

}
}
\seealso{
stabsel
}

